+++ R/importWAQN.R
@@ -32,6 +32,8 @@
 ##' @param pollutant Pollutants to import. If omitted will import all pollutants
 ##'   from a site. To import only NOx and NO2 for example use \code{pollutant =
 ##'   c("nox", "no2")}.
+##' @param meta Should meta data be returned? If \code{TRUE} the site type,
+##'   latitude and longitude are returned.
 ##' @param to_narrow By default the returned data has a column for each
 ##'   pollutant/variable. When \code{to_narrow = TRUE} the data are stacked into
 ##'   a narrow format with a column identifying the pollutant name.  
@@ -55,7 +57,9 @@
 ##' \dontrun{all <- importWAQN(site = site = c("card", "cae6"), year = 2018)}
 ##'
 ##' 
-importWAQN <- function(site = "card", year = 2018, pollutant = "all", to_narrow = FALSE) {
+importWAQN <- function(site = "card", year = 2018, pollutant = "all", 
+                       meta = FALSE,
+                       to_narrow = FALSE) {
   site <- toupper(site)
 
 
@@ -140,10 +144,26 @@ importWAQN <- function(site = "card", year = 2018, pollutant = "all", to_narrow 
   ## make sure it is in GMT
   attr(thedata$date, "tzone") <- "GMT"
   
+  if (meta) {
+    meta_data <- importMeta(source = "waqn")
+    # suppress warnings about factors
+    thedata <- suppressWarnings(inner_join(thedata, meta_data, by = c("code", "site")))
+  }
+  
   if (to_narrow) {
     
-    thedata <- pivot_longer(thedata, -c(date, site, code), names_to = "pollutant") %>% 
-      arrange(site, code, pollutant, date)
+    if (meta) {
+      
+      thedata <- pivot_longer(thedata, -c(date, site, code, latitude, longitude, site.type), 
+                              names_to = "pollutant") %>% 
+        arrange(site, code, pollutant, date)
+      
+    } else {
+      
+      thedata <- pivot_longer(thedata, -c(date, site, code), names_to = "pollutant") %>% 
+        arrange(site, code, pollutant, date)
+      
+    }
   }
 
   as_tibble(thedata)