+++ NEWS.md
@@ -6,6 +6,7 @@
 - fix issue with DST in `cutData`
 - allow 'hemisphere' to be supplied as an argument to `timeAverage`, used for avg.time = "season". 
 - add option `to_narrow` to `importAURN`, `importSAQN` and `importWAQN` to stack data into a tidy format. The data are now returned as a 'tibble'
+- Allow meta data to be returned in `importSAQN` and `importWAQN`.
 
 ## openair 2.6-4